+++ intercom_flutter/ios/Classes/IntercomFlutterPlugin.m
@@ -231,6 +231,33 @@ id unread;
     } else if([@"displayHome" isEqualToString:call.method]) {
         [Intercom presentIntercom:home];
         result(@"Presented home space");
+    } else if([@"isUserLoggedIn" isEqualToString:call.method]) {
+        if([Intercom isUserLoggedIn]) {
+            result(@(YES));
+        }else{
+            result(@(NO));
+        }
+    } else if([@"fetchLoggedInUserAttributes" isEqualToString:call.method]) {
+        ICMUserAttributes *data = [Intercom fetchLoggedInUserAttributes];
+        if(data != (id)[NSNull null]){
+            NSDictionary *attributes = data.attributes;
+            NSMutableDictionary<NSString *, id> *map = [attributes mutableCopy];
+            
+            // Add custom attributes
+            map[@"custom_attributes"] = data.customAttributes;
+           
+            // Add companies
+            if (data.companies) {
+                NSMutableArray *companiesArray = [NSMutableArray array];
+                for (ICMCompany *company in data.companies) {
+                    [companiesArray addObject:[company attributes]];
+                }
+                map[@"companies"] = companiesArray;
+            }
+            
+            result(map);
+        }
+        result([NSMutableDictionary dictionary]);
     }
     else {
         result(FlutterMethodNotImplemented);